crypto: change type uint32_t to size_t in aes_cypher() arguments.

+++ arch/risc-v/src/esp32c3/esp32c3_aes.c
@@ -587,8 +587,8 @@ int esp32c3_aes_init(void)
 
 #ifdef CONFIG_CRYPTO_AES
 
-int aes_cypher(void *out, const void *in, uint32_t size,
-               const void *iv, const void *key, uint32_t keysize,
+int aes_cypher(void *out, const void *in, size_t size,
+               const void *iv, const void *key, size_t keysize,
                int mode, int encrypt)
 {
   int ret;
